requirements.txt
celery==5.2.0
Django==3.2.9
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2-binary==2.9.2
python-decouple==3.5
pytz==2021.3
django-filter==21.1
